requirements.txt
Django==4.2.23
asgiref==3.9.1
sqlparse==0.5.3
typing_extensions==4.15.0
djangorestframework==3.14.0
django-cors-headers==4.1.0
openpyxl==3.1.2
